+++ add.php
@@ -6,7 +6,6 @@ $title = 'Добавление лота';
 
 require_once ('functions/template.php');
 require_once ('functions/db.php');
-require_once ('functions/validators.php');
 require_once ('functions/validators_lot.php');
 $config = require 'config.php';
