+++ functions/validators_user.php
@@ -96,3 +96,25 @@ function validate_user_contacts ($contacts) {
     }
     return null;
 }
+
+/**
+ * Функция проверяет валидность файла, файл должен быть изображением.
+ *
+ * @param $img_name array массив _FILES
+ *
+ * @return array
+ */
+function validate_file($img_name) {
+    $errors = [];
+    if (!empty($img_name))  {
+        $finfo     = finfo_open(FILEINFO_MIME_TYPE);
+        $file_type = finfo_file($finfo, $img_name);
+
+        if (mb_strpos ($file_type, 'image') === false) {
+            $errors['img'] = 'Загрузите картинку';
+            return $errors;
+        }
+        return $errors;
+    }
+    return $errors;
+}
