+++ lot.php
@@ -6,7 +6,7 @@ $title = 'Лот';
 
 require_once ('functions/template.php');
 require_once ('functions/db.php');
-require_once ('functions/validators.php');
+require_once ('functions/validators_lot.php');
 $config = require 'config.php';
 
 $link = db_connect($config['db']);
