+++ sign-up.php
@@ -6,7 +6,6 @@ $title = 'Регистрация нового аккаунта';
 
 require_once ('functions/template.php');
 require_once ('functions/db.php');
-require_once ('functions/validators.php');
 require_once ('functions/validators_user.php');
 $config = require 'config.php';
